vm-provisioning/requirements.txt
# SPDX-FileCopyrightText: (C) 2025 Intel Corporation
# SPDX-License-Identifier: Apache-2.0

# YAML linter
yamllint~=1.29.0

# license checking
python-debian==0.1.44
reuse~=1.0.0

PyYAML>=5.3

